% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_country.R
\name{assign_country}
\alias{assign_country}
\title{assign_country}
\usage{
assign_country(inp_df)
}
\arguments{
\item{inp_df}{the input dataframe retrieved from the fireball API}
}
\value{
the input dataframe with "country" (country name) column appended
}
\description{
Assign country name to coordinates of fireballs in a dataframe
}
