% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_country.R
\name{clean_latlon}
\alias{clean_latlon}
\title{clean_latlon}
\usage{
clean_latlon(x, direction)
}
\arguments{
\item{x}{coordinate point to process (can be latitude or longitude)}

\item{direction}{N/S for Latitude, E/W for Longitude}
}
\value{
coordinate point with correct sign applied relative to equator (south=negative) or prime meridan (west=negative)
}
\description{
Clean Latitude or Longitude values
}
