% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_country.R
\name{find_country}
\alias{find_country}
\title{find_country}
\usage{
find_country(x, y)
}
\arguments{
\item{x}{longitude point (expected as numeric)}

\item{y}{latitude point (expected as numeric)}
}
\value{
character object representing country name (i.e. "Canada") or the ocean ("OCEAN"), or an unknown location ("UNKNOWN")
}
\description{
INTERNAL FUNCTION: Helper function for finding country of coordinate pair based on polygons
}
