% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldmap_planar}
\alias{worldmap_planar}
\title{World Map Planar dataset}
\format{THIS IS PLACEHOLDER DATA!!!!!!:
\describe{

  ...
}}
\source{
\url{http://www.google.com}
}
\usage{
worldmap_planar
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
