% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireNASA.R
\name{fireball_data}
\alias{fireball_data}
\title{fireball_data}
\usage{
fireball_data(date_min, date_max, min_energy, max_energy, min_velocity,
     max_velocity, lim)
}
\arguments{
\item{date_min}{Exclude data earlier than this date YYYY-MM-DD or date/time YYYY-MM-DDThh:mm:ss}

\item{date_max}{Exclude data later than this date YYYY-MM-DD or date/time YYYY-MM-DDThh:mm:ss}

\item{min_energy}{Exclude data with Total-Radiated-Energy less than this positive value in joules (e.g.,0.3)}

\item{max_energy}{Exclude data with Total-Radiated-Energy less than this positive value}

\item{min_velocity}{Exclude data with velocity-at-peak-brightness less than this positive value in km/s (e.g., 18.5)}

\item{max_velocity}{Exclude data with velocity-at-peak-brightness greater than this positive value in km/s (e.g., 20)}

\item{lim}{Number of rows to be displayed}
}
\description{
Wrapper for downloading NASA's Fireball Data
}
\examples{
fireball_data(date_min = "2015-01-01", min_energy = 0.6, lim = 10)
}
