% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireNASA.R
\name{fireball_data}
\alias{fireball_data}
\title{fireball_data}
\usage{
fireball_data(lim, min_e = NULL)
}
\arguments{
\item{lim}{Number of Rows needed}

\item{min_e}{Exclude Data with Total-Radiated-Energy less than this positive value}
}
\description{
Wrapper for downloading NASA's Fireball Data
}
\examples{
fireball_data(lim = 10, min_e = 0.6)
}
