% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldmap_planar}
\alias{worldmap_planar}
\title{World Map Planar dataset}
\format{A Simple Features dataframe with 177 rows and 95 variables
\describe{

  ...
}}
\source{
\url{http://www.naturalearthdata.com}
}
\usage{
worldmap_planar
}
\description{
A dataset of world map shapefile data that has been transformed to planar using EPSG 2163 (US National Atlas Equal Area).
Original data is ne_110m_admin_0_countries.shp which is "Made with Natural Earth. Free vector and raster map data @ naturalearthdata.com."
}
\keyword{datasets}
